File structure cleaned up
