+++ lib/slackattend/current_member.rb
@@ -1,0 +1,6 @@
+module Slackattend
+  class Member < ActiveRecord::Base
+    validates :name, uniqueness: true
+    validates :avatar, uniqueness: true
+  end
+end
