+++ lib/slackattend/status_log.rb
@@ -1,0 +1,4 @@
+module Slackattend
+  class Status < ActiveRecord::Base
+  end
+end
